Move implementation and test into own files
